requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
decorator==4.4.2
docutils==0.21.2
edge-tts==7.0.0
frozenlist==1.5.0
gTTS==2.5.4
h11==0.14.0
id==1.5.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
loguru==0.7.3
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.6.0
moviepy==2.1.2
multidict==6.1.0
nh3==0.2.20
numpy==2.2.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
PictureTextCreator==0.2.1
pillow==10.4.0
proglog==0.1.10
propcache==0.3.0
pycparser==2.22
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32-ctypes==0.2.3
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
selenium==4.27.1
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
srt==3.5.3
tabulate==0.9.0
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
twine==6.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
wheel==0.45.1
win32_setctime==1.2.0
wsproto==1.2.0
yarl==1.18.3
